requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
pydantic==2.7.1
pytest==8.2.0
pytest-django==4.8.0
ruff==0.4.2
black==24.4.2
django-stubs[compatible-mypy]==5.0.0
